requirements.txt
Flask==3.0.0
biopython==1.83
svgwrite==1.4.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
